Add mobile live scores modal, fix Pre-Game text overflow
On mobile, move live scores into a separate pop-up modal (matching
the existing AP rankings pattern) instead of an inline scroll bar,
which was crowding the page. Extract the fetching logic into a
useLiveScores hook and the game tile into a LiveGameCard component so
the desktop bar and the new modal share the same code.

While extracting the tile, give the status label ("Pre-Game", etc.)
the full width of the box with nowrap so it no longer spills outside
its boundaries, and fix a copy/paste bug where the home team's
fallback name showed the away team's name instead.
